+++ cogs/ddo.py
@@ -22,24 +22,32 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from asyncio import sleep, wait_for, TimeoutError
-from functools import reduce
+import re
+from asyncio import sleep, TimeoutError
+from contextlib import suppress
 from json.decoder import JSONDecodeError
-from random import seed, shuffle, randrange
-from re import search, findall
+from random import seed, randrange
+from re import search
 from time import time
-from typing import List, no_type_check
+from typing import no_type_check, Dict, Optional, List
 
+import discord
 from aiohttp import ClientError
 from bs4 import BeautifulSoup
 from discord import Embed
+from discord import app_commands, Interaction
+from discord.app_commands import Choice
+from discord.app_commands.transformers import Range
 from discord.ext import commands, tasks
 
 from dreambot import DreamBot
 from utils.context import Context
+from utils.enums.ddo_enums import Server, AdventureType, Difficulty
 from utils.enums.network_return_type import NetworkReturnType
+from utils.intermediate_models.ddo_audit_models import DDOAuditServer, DDOAuditGroup, DDOAdventureEmbed, DDOPartyEmbed
 from utils.logging_formatter import bot_logger
 from utils.network_utils import network_request, ExponentialBackoff
+from utils.utils import calculate_padding
 
 
 class DDO(commands.Cog):
@@ -51,6 +59,7 @@ class DDO(commands.Cog):
         ADVENTURE_TYPES (tuple): A tuple of valid adventure types for DDOAudit
         DIFFICULTIES (tuple): A tuple of valid difficulties for DDOAudit
         QUERY_INTERVAL (int): How frequently API data from DDOAudit should be queried.
+        feature_subgroup (app_commands.Group): The AppCommand Group for commands (not implicit - `roll` is top-level)
 
     Attributes:
         bot (DreamBot): The Discord bot.
@@ -59,10 +68,13 @@ class DDO(commands.Cog):
         backoff (ExponentialBackoff): Exponential Backoff calculator for network requests.
     """
 
-    SERVERS = ('Argonnessen', 'Cannith', 'Ghallanda', 'Khyber', 'Orien', 'Sarlona', 'Thelanis', 'Wayfinder', 'Hardcore')
-    QUEST_TYPES = ('Solo', 'Quest', 'Raid')
-    DIFFICULTIES = ('Casual', 'Normal', 'Hard', 'Elite', 'Reaper')
     QUERY_INTERVAL = 15
+
+    ddo_subgroup = app_commands.Group(
+        name='ddo',
+        description='Commands for Dungeons & Dragons Online',
+        guild_only=False
+    )
 
     def __init__(self, bot: DreamBot) -> None:
         """
@@ -73,11 +85,12 @@ class DDO(commands.Cog):
         """
 
         self.bot = bot
-        self.api_data = {server: None for server in self.SERVERS}
+        self.api_data: Dict[Server, Optional[DDOAuditServer]] = {server: None for server in Server}
         self.backoff = ExponentialBackoff(60 * 60 * 4)
+        self.roll_regex = re.compile('(?P<quantity>\d{0,6})d(?P<sides>\d{1,6}) ?(?P<modifier>[-+] ?\d{1,6})?')
         self.query_ddo_audit.start()
 
-    @commands.command(name='roll', help='Simulates rolling dice. Syntax example: 9d6')
+    @commands.hybrid_command(name='roll', help='Simulates rolling dice. Syntax example: 9d6')  # type: ignore[arg-type]
     async def roll(self, ctx: Context, *, pattern: str) -> None:
         """
         A method to simulate the rolling of dice.
@@ -86,100 +99,40 @@ class DDO(commands.Cog):
             ctx (Context): The invocation context.
             pattern (str): The die pattern to roll. Example: 9d6 -> Nine Six-Sided die.
 
-        Output:
-            Success: The result of the die pattern rolled.
-            Failure: A description of the syntax error that occurred.
-
         Returns:
             None.
         """
 
-        async def evaluate_dice(dice_pattern: str) -> List[int]:
-            """
-            Simulate rolling the specified dice.
+        match = self.roll_regex.match(pattern)
 
-            Parameters:
-                dice_pattern (str): The specified dice pattern to generate.
+        if match is None:
+            raise commands.UserInputError('Could not extract a valid dice roll pattern.')
 
-            Returns:
-                (List[int]): The result of rolling the specified pattern.
-            """
+        die_quantity = match.group('quantity')
+        die_sides = match.group('sides')
+        modifier = match.group('modifier')
 
-            die = dice_pattern.split('d')
-            # create a list of all possible roll outcomes
-            # dice = [x for x in range(int(die[0]), int(die[0]) * int(die[1]) + 1)]
-            dice = []
-            for _ in range(10):
-                seed(time())
-                dice.append([randrange(1, int(die[1]) + 1) for _ in range(int(die[0]))])
-            shuffle(dice)
-            # shuffle the list and send the first index's result
-            return dice[0]
+        # sides is required, quantity and modifier are optional
+        die_sides = parse_match(die_sides, 'Die Sides')
 
-        async def evaluate_dice_string(die_string: str) -> str:
-            """
-            Turns a die string into a computed value without exposing the string to eval.
+        try:
+            die_quantity = parse_match(die_quantity, 'Number of Dice')
+        except commands.UserInputError:
+            die_quantity = 1
 
-            Parameters:
-                die_string (str): The die string to be evaluated.
+        try:
+            modifier = parse_match(modifier, 'Modifier', negative_allowed=True)
+        except commands.UserInputError:
+            modifier = 0
 
-            Returns:
-                (str): The evaluated result.
-            """
+        seed(time())
+        die_total = sum(randrange(1, die_sides + 1) for _ in range(die_quantity)) + modifier
 
-            # avoid exposing eval() to the user -> manually parse the arithmetic expression we've generated
-            # while we have valid expressions, break them down into groups
-            while match := search(r'(\d+)([+\-])(\d+)', die_string):
-                groups = match.groups()
-
-                if groups[1] == '+':
-                    total = int(groups[0]) + int(groups[2])
-                else:
-                    total = int(groups[0]) - int(groups[2])
-
-                # replace the expression with the result and continue
-                die_string = die_string.replace(f'{"".join(groups)}', str(total), 1)
-
-            return die_string
-
-        # remove all spaces from the string, and manually add a space to the front
-        # this allows this regex pattern to find a 'd#' at the beginning
-        pattern = ' ' + pattern.replace(' ', '')
-        # build a dict of the single die in the pattern (ex: 'd20', 'd2', etc.)
-        single_die = {x[0] + x[1]: x[0] + '1' + x[1] for x in findall(r'(\D)(d\d+)', pattern)}
-        # replace all the single die with a '1d#' alternative, ensuring all dice follow the #d# format
-        for key, value in single_die.items():
-            pattern = pattern.replace(key, value.strip(), 1)
-        # with all die in the same format, extract all the requested rolls
-        die_patterns = findall(r'\d+d\d+', pattern)
-        # build a dict of results {request: result}
-        results = {die: await evaluate_dice(die) for die in die_patterns}
-        # build a result string we can present to the user
-        breakdown = f'Roll: **{pattern}**\nResult: **$**\n\nBreakdown:'
-
-        # group all the non-die rolls together, so we can append it the breakdown
-        non_die_base = reduce(lambda s, r: s.replace(r, '@', 1), die_patterns, pattern)
-        non_die = findall(r'[+|-]\d+|\d+(?=[+|-])', non_die_base)
-
-        # give the user a breakdown of each of their requested rolls
-        for key, value in results.items():
-            pattern = pattern.replace(key, str(sum(value)), 1)
-            breakdown += f'\n{key} ({sum(value)}): {value}'
-
-        # if the user supplied non-die args, add those to the breakdown
-        if non_die:
-            non_die_total = sum([int(x) for x in non_die])
-            breakdown += f'\nNon-Die ({non_die_total}): {[int(x) for x in non_die]}'
-
-        async with ctx.channel.typing():
-            try:
-                final_value = await wait_for(evaluate_dice_string(pattern), 3)
-                await ctx.send(breakdown.replace('$', final_value))
-            except TimeoutError:
-                await ctx.send('Die Evaluation Timeout Error')
+        await ctx.send(f'{die_total:,}')
 
     # noinspection GrazieInspection
     # needs cleanup
+    # -> 6/22/24 - ignoring during migration as command is essentially disabled
     @no_type_check
     @commands.is_owner()
     @commands.command(name='ddoitem', help='Pulls basic information about an item in Dungeons & Dragons Online '
@@ -263,8 +216,8 @@ class DDO(commands.Cog):
                     # Note: This description was written 'backwards', (positive lookbehind occurs first, etc.)
                     result = search("(?<=>)( )*(\+\d+ )*(\w|\d)(.+?)(?=</a>)", element)
                     # If our result is not a Mythic Bonus (these are on nearly every single item), add it
-                    if result and str(result.group(0)).find('Mythic') == -1:
-                        enchantments.append(result.group(0).strip())
+                    if result and str(result.group()).find('Mythic') == -1:
+                        enchantments.append(result.group().strip())
 
         # Prep other detail variables
         item_type = 'none'
@@ -272,11 +225,11 @@ class DDO(commands.Cog):
 
         # If we have a minimum level string, extract the minimum level using regex
         if minimum_level_string is not None:
-            minimum_level = int(search("\d+?(?=(\n</td>))", minimum_level_string).group(0))
+            minimum_level = int(search("\d+?(?=(\n</td>))", minimum_level_string).group())
 
         # If we have an item type string, extract the item type using regex
         if item_type_string is not None:
-            item_type = search("(?<=>).+?(?=(\n</td>))", item_type_string).group(0)
+            item_type = search("(?<=>).+?(?=(\n</td>))", item_type_string).group()
             # Sometimes (in the case of weapons) the parent type is bolded - strip these tags
             item_type = item_type.replace('<b>', '').replace('</b>', '')
 
@@ -297,140 +250,96 @@ class DDO(commands.Cog):
         embed.set_footer(text="Please report any formatting issues to my owner!")
         await ctx.send(embed=embed)
 
-    @commands.command(name='lfms', help=f'Returns a list of active LFMs for the specified server.\nValid servers'
-                                        f' include Argonnessen, Cannith, Ghallanda, Khyber, Orien, Sarlona, Thelanis,'
-                                        f' and Wayfinder\nInformation is populated from \'DDO Audit\' every '
-                                        f'{QUERY_INTERVAL} seconds.')
-    async def ddo_lfms(self, ctx: Context, server: str = 'Khyber') -> None:
+    @ddo_subgroup.command(
+        name='lfms',
+        description=f'Returns active LFMs on the specified server. '
+                    f'Data is populated from \'DDO Audit\' every {QUERY_INTERVAL} seconds.'
+    )
+    @app_commands.describe(
+        server='The server to show LFMs for',
+        adventure_type='Optional: Filter groups by adventure type',
+        raw_difficulty='Optional: Filter groups by difficulty',
+        level='Optional: Filter groups by supported level'
+    )
+    @app_commands.choices(raw_difficulty=Difficulty.choices())
+    @app_commands.rename(raw_difficulty='difficulty')
+    async def ddo_lfms(
+            self,
+            interaction: Interaction[DreamBot],
+            server: Server,
+            raw_difficulty: Optional[Choice[str]] = None,
+            adventure_type: Optional[AdventureType] = None,
+            level: Optional[Range[int, 1, 34]] = None
+    ) -> None:
         """
         A method that outputs a list of all active groups on a server.
 
         Parameters:
-            ctx (Context): The invocation context.
-            server (str): The name of the server to return lfms for. Default: 'Khyber'.
-
-        Output:
-            Success: A message detailing the lfms per category.
-            Failure: A description of the error that occurred.
-
-        Returns:
-            None.
-        """
-
-        if server not in self.SERVERS:
-            server = 'Khyber'
-
-        try:
-            server_data = self.api_data[server]
-
-            if server_data is None:
-                raise ValueError
-
-        except KeyError:
-            await ctx.send(f'No Active LFM\'s on {server}!')
-            return
-        except ValueError:
-            await ctx.send('Failed to query DDO Audit API.')
-            return
-
-        # Divide the groups into three lists: Raids, Quests, and Groups (No Listed Quest)
-        else:
-            raids = [q['Quest']['Name'] for q in server_data['Groups'] if
-                     q['Quest'] and q['Quest']['GroupSize'] == 'Raid']
-            quests = [q['Quest']['Name'] for q in server_data['Groups'] if
-                      q['Quest'] and q['Quest']['GroupSize'] != 'Raid']
-            groups = [q['Comment'] for q in server_data['Groups'] if not q['Quest'] and q['Comment']]
-
-            # Should a list be empty, append 'None'
-            for li in [raids, quests, groups]:
-                if not li:
-                    li.append('None')
-
-            await ctx.send(f'**Current Raids on {server}:** {", ".join(raids)}\n'
-                           f'**Current Quests on {server}:** {", ".join(quests)}\n'
-                           f'**Current Groups on {server}:** {", ".join(groups)}\n')
-
-    @commands.command(name='flfms', help=f'Returns a filtered list of active LFMs for the specified server.\n'
-                                         f'Optional filters include: LFM Type: (Solo, Quest, Raid), Difficulty: '
-                                         f'(Casual, Normal, Hard, Elite, Reaper), and Level: (1-32).\nYou MUST supply a'
-                                         f' server.\nValid servers include Argonnessen, Cannith, Ghallanda, Khyber,'
-                                         f' Orien, Sarlona, Thelanis, Wayfinder, and Hardcore.\nInformation is '
-                                         f'populated from \'DDO Audit\' every {QUERY_INTERVAL} seconds.')
-    async def ddo_filter_lfms(self, ctx: Context, *args: str) -> None:
-        """
-        A method that outputs a list of all active groups on a server that match the specified filters.
-
-        Parameters:
-            ctx (Context): The invocation context.
-            args (str): The filter options. Options include Type, Difficulty, Level, and a non-optional Server.
-
-        Output:
-            Success: A message detailing the filtered lfms.
-            Failure: A description syntax error that occurred.
+            interaction (Interaction): The invocation interaction.
+            server (Server): The name of the server to return lfms for.
+            adventure_type (Optional[AdventureType]): Filter groups by this adventure type.
+            raw_difficulty (Optional[Choice[str]]): Filter groups by this difficulty.
+            level (Optional[int]): Filter groups by this level.
 
         Returns:
             None.
         """
 
-        # attempt to parse arguments provided
-        server = atype = diff = level = None
+        server_data: Optional[DDOAuditServer] = self.api_data[server]
+        difficulty: Optional[Difficulty] = Difficulty.from_string(raw_difficulty.value) if raw_difficulty else None
 
-        for arg in args:
-            # parse string arguments first
-            if arg in self.SERVERS:
-                server = arg
-            elif arg in self.QUEST_TYPES:
-                atype = arg
-            elif arg in self.DIFFICULTIES:
-                diff = arg
-            # if we can't parse the arg to one of the above, try to convert the arg to an int
-            # if we can't cast to int OR a successful cast is outside our acceptable range, level = None
-            else:
-                try:
-                    level = int(arg)
-                    if level < 1 or level > 32:
-                        level = None
-                except ValueError:
-                    level = None
-
-        if server not in self.SERVERS:
-            server = 'Khyber'
-
-        try:
-            server_data = self.api_data[server]
-
-            if server_data is None:
-                raise ValueError
-
-        except KeyError:
-            await ctx.send(f'No Active LFM\'s on {server}!')
-            return
-        except ValueError:
-            await ctx.send('Failed to query DDO Audit API.')
+        if server_data is None:
+            await interaction.response.send_message(
+                f'Server data for {server} is not currently available.', ephemeral=True
+            )
             return
 
-        # build sets for each of our individual filters, as well as a master set of all quests
-        # sets are tuples of (LeaderName, QuestName, Difficulty, AdventureType), with LeaderName
-        #   included to allow for different hashes of otherwise identical groups
+        if server_data.group_count is None or server_data.group_count == 0:
+            await interaction.response.send_message(f'No parties currently available on {server}.', ephemeral=True)
+            return
 
-        all_quests = {(q['Leader']['Name'], q['Quest']['Name'], q['Difficulty'], q['Quest']['GroupSize'])
-                      for q in server_data['Groups'] if q['Quest']}
-        atypes = {(q['Leader']['Name'], q['Quest']['Name'], q['Difficulty'], q['Quest']['GroupSize'])
-                  for q in server_data['Groups'] if q['Quest'] and atype and q['Quest']['GroupSize'] == atype}
-        diffs = {(q['Leader']['Name'], q['Quest']['Name'], q['Difficulty'], q['Quest']['GroupSize'])
-                 for q in server_data['Groups'] if q['Quest'] and diff and q['Difficulty'] == diff}
-        levels = {(q['Leader']['Name'], q['Quest']['Name'], q['Difficulty'], q['Quest']['GroupSize']) for q
-                  in server_data['Groups'] if q['Quest'] and level and q['MinimumLevel'] <= level <= q['MaximumLevel']}
+        base_groups: List[DDOAuditGroup] = server_data.groups
 
-        # if our value is not None, start performing intersection calculations on the full set
-        for filtered_set, value in [(atypes, atype), (diffs, diff), (levels, level)]:
-            if value is not None:
-                all_quests.intersection_update(filtered_set)
+        # -- filtering --
+        filters: List[str] = []
 
-        if not all_quests:
-            await ctx.send(f'**Filtered Results on {server}:** None')
-        else:
-            await ctx.send(f'**Filtered Results on {server}:** {", ".join(x[1] for x in all_quests)}')
+        if adventure_type is not None:
+            base_groups = [x for x in base_groups if x.quest and x.quest.group_size == adventure_type.ddo_audit_value]
+            filters.append(adventure_type.value)
+
+        if difficulty is not None:
+            base_groups = [x for x in base_groups if x.difficulty in difficulty.difficulty_set]
+            filters.append(difficulty.value)
+
+        if level is not None:
+            base_groups = [
+                x for x in base_groups
+                if x.minimum_level is not None and x.maximum_level is not None
+                   and x.minimum_level <= level <= x.maximum_level
+            ]
+            filters.append(f'Level {level}')
+
+        # -- splitting --
+        quests: List[DDOAuditGroup] = [
+            group for group in base_groups if group.quest and group.quest.group_size == 'Party'
+        ]
+        raids: List[DDOAuditGroup] = [
+            group for group in base_groups if group.quest and group.quest.group_size == 'Raid'
+        ]
+        parties: List[DDOAuditGroup] = [
+            group for group in base_groups if
+            (group.quest is None and group.comment is not None) or
+            (group.guess == True and group.comment is not None)
+        ]
+
+        await send_lfms_response(
+            interaction,
+            server,
+            filters if filters else None,
+            quests[::-1],
+            raids[::-1],
+            parties[::-1]
+        )
 
     @tasks.loop(seconds=QUERY_INTERVAL)
     async def query_ddo_audit(self) -> None:
@@ -445,7 +354,7 @@ class DDO(commands.Cog):
             None.
         """
 
-        async def backoff(current_server: str) -> None:
+        async def backoff(current_server: Server) -> None:
             """
             A local function for handling exponential backoff during network errors.
             Backoff behavior is desirable during any exception, though other behavior is exception-specific.
@@ -463,7 +372,7 @@ class DDO(commands.Cog):
             self.backoff.next_backoff()
 
             if self.backoff.backoff_count >= 4:
-                self.api_data = {s: None for s in self.SERVERS}
+                self.api_data = {s: None for s in Server}
                 bot_logger.warning(
                     f'DDOAudit Total Backoff Duration Will Exceed 5 Minutes. '
                     f'Clearing all LFM data and backing off for {self.backoff.str_time}.'
@@ -473,14 +382,15 @@ class DDO(commands.Cog):
 
             await sleep(self.backoff.total_backoff_seconds)
 
-        server = self.SERVERS[self.query_ddo_audit.current_loop % len(self.SERVERS)]
+        server = list(Server)[self.query_ddo_audit.current_loop % len(Server)]
 
         try:
-            self.api_data[server] = await network_request(
+            server_data = await network_request(
                 self.bot.session,
-                f'https://api.ddoaudit.com/groups/{server.lower()}',
+                f'https://api.ddoaudit.com/groups/{server.value.lower()}',
                 return_type=NetworkReturnType.JSON, ssl=False
             )
+            self.api_data[server] = DDOAuditServer(server_data)
             self.backoff.reset()
 
         except ClientError:
@@ -508,9 +418,120 @@ class DDO(commands.Cog):
         """
 
         self.query_ddo_audit.cancel()
-        self.api_data = dict()
 
         bot_logger.info('Completed Unload for Cog: DDO')
+
+
+async def send_lfms_response(
+        interaction: Interaction[DreamBot],
+        server: Server,
+        filters: Optional[List[str]],
+        quests: List[DDOAuditGroup],
+        raids: List[DDOAuditGroup],
+        parties: List[DDOAuditGroup]
+) -> None:
+    """
+    Generates an embed detailing Group results for the /lfm command.
+
+    Parameters:
+        interaction (Interaction): The invocation interaction.
+        server (Server): The specified server to display LFMs for.
+        filters (Optional[List[str]]): The filters used to refine results, if any.
+        quests (List[DDOAuditGroup]): The resulting list of groups in the 'Quest' category.
+        raids (List[DDOAuditGroup]): The resulting list of groups in the 'Raid' category.
+        parties (List[DDOAuditGroup): The resulting list of groups in that do not specify an adventure.
+
+    Returns:
+        None.
+    """
+
+    # group count should always be non-zero at this point since we've already early returned if == 0
+    result_count = len(quests) + len(raids) + len(parties)
+
+    if result_count == 0 and filters is not None:
+        await interaction.response.send_message('No parties current match the specified filters.')
+        return
+
+    response: str = f'**__DDO Parties on {server}__**\n'
+
+    if filters is not None:
+        response += (f'_{result_count:,} part{"ies" if result_count != 1 else "y"} currently '
+                     f'match{"" if result_count != 1 else "es"} the '
+                     f'filter{"s" if len(filters) != 1 else ""}!'
+                     f'\nFilter{"s" if len(filters) != 1 else ""}: {", ".join(filters)}_')
+    else:
+        response += f'_{result_count:,} part{"ies" if result_count != 1 else "y"} currently available!_'
+
+    sanitized_quests = [embed_component for x in quests if (embed_component := DDOAdventureEmbed.from_group(x))]
+    sanitized_raids = [embed_component for x in raids if (embed_component := DDOAdventureEmbed.from_group(x))]
+    sanitized_parties = [embed_component for x in parties if (embed_component := DDOPartyEmbed.from_group(x))]
+
+    response_components: List[str] = []
+
+    if sanitized_quests:
+        title_padding = calculate_padding(sanitized_quests, 'title')
+        difficulty_padding = calculate_padding(sanitized_quests, 'difficulty')
+        members_padding = calculate_padding(sanitized_quests, 'members_description')
+
+        embed_description = '\n'.join(
+            x.full_description(title_padding, difficulty_padding, members_padding) for x in sanitized_quests
+        )
+
+        response_components.append(f'__*Quests*__\n```\n{embed_description}```')
+
+    if sanitized_raids:
+        title_padding = calculate_padding(sanitized_raids, 'title')
+        difficulty_padding = calculate_padding(sanitized_raids, 'difficulty')
+        members_padding = calculate_padding(sanitized_raids, 'members_description')
+
+        embed_description = '\n'.join(
+            x.full_description(title_padding, difficulty_padding, members_padding) for x in sanitized_raids
+        )
+
+        response_components.append(f'__*Raids*__\n```\n{embed_description}```')
+
+    if sanitized_parties:
+        title_padding = calculate_padding(sanitized_parties, 'title')
+        members_padding = calculate_padding(sanitized_parties, 'members_description')
+
+        embed_description = '\n'.join(
+            x.full_description(title_padding, 0, members_padding) for x in sanitized_parties
+        )
+
+        response_components.append(f'__*Groups*__\n```\n{embed_description}```')
+
+    response += '\n\n' + '\n'.join(response_components)
+
+    with suppress(discord.HTTPException):
+        # TODO: needs a 2k character check.. batch send. Edge case for now, though.. unlikely to exceed 2k normally.
+        await interaction.response.send_message(response)
+
+
+def parse_match(match: str, match_name: str, negative_allowed: bool = False) -> int:
+    """
+    Parses a match string to a valid integer value.
+
+    Parameters:
+        match (str): The match string.
+        match_name (str): The name of match being parsed.
+        negative_allowed (bool): Whether negative values are allowed.
+
+    Throws:
+        commands.UserInputError.
+
+    Returns:
+        (int).
+    """
+
+    try:
+        base_value = int(match.replace(' ', ''))
+    except (ValueError, AttributeError):
+        raise commands.UserInputError(f'Could not parse {match_name}.')
+
+    if not negative_allowed and base_value <= 0:
+        raise commands.UserInputError(f'{match_name} was out-of-bounds (negative or zero).')
+
+    return base_value
 
 
 async def setup(bot: DreamBot) -> None: